Split consumer poll tests
